Merged PR 15948: Fix image url when version is provided
Fix image url when version is provided

The image url was not properly formed for firmware version: 202505.
Updated the case to include 202505 versions when creating the image url.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to correctly update the image URL suffix when a particular version is provided.

#### PR Summary
This pull request updates the pipeline configuration to address a versioning issue in image URLs.
- In `/.azure-pipelines/sonic-metadata/upgrade_sonic_metadata.yml`, a new conditional branch is added to set `IMAGE_SUFFIX` to `20250510` when the image URL includes "202505".
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->