Merged PR 16346: Change default upgrade iterations to 22 on tbtk5-t0-2700-1 pipeline
Change default upgrade iterations to 22 on tbtk5-t0-2700-1 pipeline

The default iterations with count 27 was causing longer time to
complete, it was cuasing the testbed to be locked when the next cron job
was being triggered.

ADO: https://msazure.visualstudio.com/One/_workitems/edit/35036871

----
#### AI description  (iteration 1)
#### PR Classification
Pipeline configuration update.

#### PR Summary
This pull request updates the pipeline configuration for the tbtk5-t0-2700-1 by changing the default upgrade iterations from "27" to "22".
- `/.azure-pipelines/sonic-metadata/mellanox/tbtk5-t0-2700-1.yml`: Modified the default value of the ITERATIONS parameter.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->