Merged PR 17018: Remove multi sad for 7260 to improve upgrade path time
Please review this PR to remove the multi_sad case for the 7260 platform in order to reduce overall upgrade path duration. The attached screenshot shows that subsequent runs are frequently blocked due to the extended time required by the SAD cases.

![image.png](https://dev.azure.com/mssonic/12b9cbf4-b1d3-4768-8e49-669345c32e5d/_apis/git/repositories/5380e8f7-6e2a-4154-8dee-f3be7b096894/pullRequests/17018/attachments/image.png)

----
#### AI description  (iteration 1)
#### PR Classification
This pull request updates the pipeline configuration to remove the "multi_sad" test case, improving the upgrade path execution time.

#### PR Summary
The changes streamline the SAD test cases for the upgrade path by eliminating the "multi_sad" case in the pipeline YAML.
- In `/.azure-pipelines/sonic-metadata/arista/vms24-t0-7260-2.yml`, the default value for `SELECT_SAD_CASES` is updated to remove "multi_sad".
- A minor adjustment is made to the `MULTI_HOP_BASE_UPGRADE_PATH` parameter configuration to maintain consistency.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->